refactor: adapt fixed queue
